[etc] chat파일구조 변경
