+++ package.json
@@ -13,6 +13,7 @@
     "@react-navigation/bottom-tabs": "^7.3.10",
     "@react-navigation/native": "^7.1.8",
     "@react-navigation/native-stack": "^7.3.12",
+    "@react-navigation/stack": "^7.3.1",
     "babel-plugin-module-resolver": "^5.0.2",
     "nativewind": "^4.1.23",
     "react": "19.0.0",
