--- src/hook/useNotificationNavigation.tsx
@@ -1,10 +1,0 @@
-// import { useNavigation } from '@react-navigation/native';
-// import { NativeStackNavigationProp } from '@react-navigation/native-stack';
-//  import { NotificationStackParamList } from '@/Navigation/NotificationNavigator';
-
-
-// type Navigation = NativeStackNavigationProp<NotificationStackParamList>;
-
-// export default function useNotificationNavigation() {
-//   return useNavigation<Navigation>();
-// }
